refactor(feature): move the feature flag types into common/feature
TitanFeatures and SingletonFeatureManagerProvider stayed in common/utils
when common/feature was created, and a new test was added next to them
there. common/utils is the catch-all package OLF-L3-02 names as the
anti-pattern, and it lists both classes by name with common/feature as
their destination.

Move both plus TitanFeaturesTest, and point the FeatureManagerProvider
service file at the new name. Callers only change their import.

ThreadHelper stays in common/utils: it is a cross-project duplicate whose
destination is Butterfly (OLF-L2-04, open point 4), not another Titan
package. Its javadoc now says so.
