+++ app/src/main/java/net/onelitefeather/titan/app/helper/NavigationHelper.java
@@ -28,8 +28,8 @@ import net.minestom.server.item.ItemStack;
 import net.onelitefeather.deliver.DeliverComponent;
 import net.onelitefeather.titan.api.deliver.Deliver;
 import net.onelitefeather.titan.common.feature.FeatureGate;
+import net.onelitefeather.titan.common.feature.TitanFeatures;
 import net.onelitefeather.titan.common.utils.Items;
-import net.onelitefeather.titan.common.utils.TitanFeatures;
 import net.theevilreaper.aves.inventory.InventoryLayout;
 import net.theevilreaper.aves.inventory.PersonalInventoryBuilder;
 import net.theevilreaper.aves.inventory.click.ClickHolder;
