+++ app/src/test/java/net/onelitefeather/titan/app/helper/NavigationHelperTest.java
@@ -26,8 +26,8 @@ import net.minestom.testing.extension.MicrotusExtension;
 import net.onelitefeather.titan.app.testutils.DummyDeliver;
 import net.onelitefeather.titan.app.testutils.TestFeatureGate;
 import net.onelitefeather.titan.common.feature.ReleaseStage;
+import net.onelitefeather.titan.common.feature.TitanFeatures;
 import net.onelitefeather.titan.common.utils.Items;
-import net.onelitefeather.titan.common.utils.TitanFeatures;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
