+++ common/src/main/resources/META-INF/services/org.togglz.core.spi.FeatureManagerProvider
@@ -1,1 +1,1 @@
-net.onelitefeather.titan.common.utils.SingletonFeatureManagerProvider
+net.onelitefeather.titan.common.feature.SingletonFeatureManagerProvider
